add generate and validate commands
