+++ go.mod
@@ -3,7 +3,10 @@ module github.com/hasura/ddn-assets
 go 1.22.4
 
 require (
+	github.com/inconshreveable/mousetrap v1.1.0 // indirect
 	github.com/machinebox/graphql v0.2.2 // indirect
 	github.com/pkg/errors v0.9.1 // indirect
+	github.com/spf13/cobra v1.8.1 // indirect
+	github.com/spf13/pflag v1.0.5 // indirect
 	golang.org/x/sync v0.8.0 // indirect
 )
